Add itemNotFound view

+++ controllers/itemsController.js
@@ -1,5 +1,7 @@
 const db = require("../db/queries");
 
+const readline = require("readline");
+
 async function getAllItems(req, res) {
   const items = await db.getAllItems();
   return items;
@@ -18,10 +20,41 @@ async function getItemByModel(req, res) {
 }
 
 async function deleteItemById(req, res) {
-  const { id } = req.params;
-  const query = String(id);
-  await db.deleteItemById(query);
-  return;
+  const admin = process.env.ADMIN_PASS;
+  let pass = "";
+  console.log("pre");
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+
+  pass = await new Promise((resolve) => {
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+
+    rl.question("Enter password: ", (answer) => {
+      rl.close();
+      resolve(answer); // Resolving with the entered password
+    });
+  });
+
+  // await rl.question("Enter password: ", (answer) => {
+  //   pass = answer;
+  //   rl.close();
+  // });
+  console.log("post");
+  console.log(pass);
+  if (pass != admin) {
+    // window.alert("Wrong password");
+    return;
+  } else {
+    const { id } = req.params;
+    const query = String(id);
+    await db.deleteItemById(query);
+    return;
+  }
 }
 
 async function getItemByModelSearch(req, res) {
